update getAll jobs

+++ server/controllers/jobController.js
@@ -1,13 +1,8 @@
 import job from "../models/JobModel.js";
-
-import { nanoid } from "nanoid";
-
-let jobs = [
-  { id: nanoid(), company: "apple", position: "front-end" },
-  { id: nanoid(), company: "google", position: "back-end" },
-];
+import mongoose from "mongoose";
 
 export const getAllJobs = async (req, res) => {
+  const jobs = await job.find({});
   res.status(200).json({ jobs });
 };
 
@@ -24,38 +19,44 @@ export const createJob = async (req, res) => {
 
 export const getjob = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ message: `invalid job id ${id}` });
+  }
+
   const job = await job.findById(id);
   if (!job) {
-    return res.status(404).json({ msg: `no ob with id ${id}` });
+    return res.status(404).json({ msg: `no job with id ${id}` });
   }
   res.status(200).json({ job });
 };
 
 export const updateJob = async (req, res) => {
-  const { company, position } = req.body;
-  if (!company || !position) {
-    res
-      .status(400)
-      .json({ message: "please provide company name and position" });
-    return;
-  }
-
   const { id } = req.params;
-  const job = job.find((job) => job.id === id);
-  if (!job) {
+
+  const updateJob = await job.findByIdAndDelete(id, req.body, {
+    new: true,
+  });
+
+  if (!updateJob) {
     return res.status(404).json({ message: `no job with id ${id}` });
   }
-  job.company = company;
-  job.position = position;
 
-  res.status(200).json({ message: "job modified", job });
+  res.status(200).json({ message: "job modified", job: updateJob });
 };
 
 export const deleteJob = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ message: `invalid job id ${id}` });
+  }
+
   const removeJob = await job.findByIdAndDelete(id);
+  console.log(removeJob);
+
   if (!removeJob) {
-    return res.status(400).json({ message: `no job with id ${id}` });
+    return res.status(404).json({ message: `no job with id ${id}` });
   }
 
   res.status(200).json({ message: "job deleted", job: removeJob });
